# modules/dxp/apps/portal-workflow/portal-workflow-metrics-rest-impl/rest-openapi.yaml
components:
    schemas:
        Assignee:
            description:
                https://schema.org/Assignee
            properties:
                id:
                    description:
                        The user's ID.
                    format: int64
                    readOnly: true
                    type: integer
                image:
                    description:
                        A relative URL to the user's profile image.
                    format: uri
                    readOnly: true
                    type: string
                name:
                    description:
                        The user's full name.
                    readOnly: true
                    type: string
            type: object
...
        Creator:
            description:
                https://schema.org/Creator
            properties:
                id:
                    description:
                        The user's ID.
                    format: int64
                    readOnly: true
                    type: integer
                name:
                    description:
                        The user's full name.
                    readOnly: true
                    type: string
            type: object
...
        Instance:
            description:
                https://www.schema.org/Instance
            properties:
                assetTitle:
                    type: string
                assetType:
                    type: string
                assignees:
                    items:
                        $ref: "#/components/schemas/Assignee"
                    readOnly: true
                    type: array
                className:
                    type: string
                classPK:
                    format: int64
                    type: integer
                completed:
                    type: boolean
                creator:
                    $ref: "#/components/schemas/Creator"
                dateCompletion:
                    format: date-time
                    type: string
                dateCreated:
                    format: date-time
                    type: string
                dateModified:
                    format: date-time
                    type: string
                description:
                    type: string
                duration:
                    format: int64
                    type: integer
                id:
                    format: int64
                    type: integer
                instance:
                    $ref: "#/components/schemas/Instance"
                processId:
                    format: int64
                    type: integer
                processVersion:
                    type: string
                slaResults:
                    items:
                        $ref: "#/components/schemas/SLAResult"
                    readOnly: true
                    type: array
                slaStatus:
                    enum: [OnTime, Overdue, Untracked]
                    readOnly: true
                    type: string
                status:
                    enum: [Completed, Pending]
                    readOnly: true
                    type: string
                taskNames:
                    items:
                        type: string
                    readOnly: true
                    type: array
                transitions:
                    items:
                        $ref: "#/components/schemas/Transition"
                    type: array
            type: object
...
        SLAResult:
            description:
                https://www.schema.org/SLA
            properties:
                dateOverdue:
                    format: date-time
                    type: string
                id:
                    format: int64
                    type: integer
                name:
                    type: string
                onTime:
                    type: boolean
                remainingTime:
                    format: int64
                    type: integer
                status:
                    enum: [Paused, Running, Stopped]
                    type: string
            type: object
...
        Transition:
            description:
                Represents the transition to be launched by the task's workflow.
            properties:
                label:
                    readOnly: true
                    type: string
                name:
                    readOnly: true
                    type: string
            type: object
info:
    description:
        ""
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: ""
    version: v1.0
openapi: 3.0.1
paths:
...
    "/processes/{processId}/instances":
        get:
            operationId: getProcessInstancesPage
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: assigneeIds
                  schema:
                      items:
                          format: int64
                          type: integer
                      type: array
                - in: query
                  name: dateEnd
                  schema:
                      format: date-time
                      type: string
                - in: query
                  name: dateStart
                  schema:
                      format: date-time
                      type: string
                - in: query
                  name: slaStatuses
                  schema:
                      items:
                          enum: [OnTime, Overdue, Untracked]
                          type: string
                      type: array
                - in: query
                  name: statuses
                  schema:
                      items:
                          enum: [Completed, Pending]
                          type: string
                      type: array
                - in: query
                  name: taskKeys
                  schema:
                      items:
                          type: string
                      type: array
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Instance"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Instance"
                                type: array
                    description:
                        ""
            tags: ["Instance"]
        post:
            operationId: postProcessInstance
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Instance"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Instance"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Instance"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Instance"
                    description:
                        ""
            tags: ["Instance"]
    "/processes/{processId}/instances/{instanceId}":
        delete:
            operationId: deleteProcessInstance
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: instanceId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        ""
            tags: ["Instance"]
        get:
            operationId: getProcessInstance
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: instanceId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Instance"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Instance"
                    description:
                        ""
            tags: ["Instance"]
        put:
            operationId: putProcessInstance
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: instanceId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Instance"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Instance"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Instance"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Instance"
                    description:
                        ""
            tags: ["Instance"]
...